Revert "Make int to float conversion explicit"
This reverts commit d483887b2c5cba0a7d15a71b7451cc1b24ef6ae4.

Reason for revert: <INSERT REASONING HERE>

+++ av1/encoder/pass2_strategy.c
@@ -1014,19 +1014,19 @@ static void define_gf_group(AV1_COMP *cpi, FIRSTPASS_STATS *this_frame,
       (float)(avg_sr_coded_error / num_mbs),
       (float)(avg_tr_coded_error / num_mbs),
       (float)(avg_wavelet_energy / num_mbs),
-      (float)(rc->constrained_gf_group),
+      rc->constrained_gf_group,
       (float)decay_accumulator,
       (float)(first_frame_coded_error / num_mbs),
       (float)(first_frame_sr_coded_error / num_mbs),
       (float)(first_frame_tr_coded_error / num_mbs),
       (float)(gf_first_frame_err / num_mbs),
-      (float)(twopass->kf_zeromotion_pct),
+      twopass->kf_zeromotion_pct,
       (float)(last_frame_coded_error / num_mbs),
       (float)(last_frame_sr_coded_error / num_mbs),
       (float)(last_frame_tr_coded_error / num_mbs),
-      (float)i,
+      i,
       (float)mv_ratio_accumulator,
-      (float)non_zero_stdev_count
+      non_zero_stdev_count
     };
     /* clang-format on */
     // Infer using ML model.
